Cleaned up failures for GET requests...
We can't do it for POSTs because we can't catch them with WebView or WebViewClient. We will have to come up with a solution to this.

Added a spinner when the page is loading for a bit of visual feed back.

This will be the first released version of the Android app.