Allows pinch zoom on full screen images.
Also to note. Due an issue with my computer I have had to change the
package name from `droidotter` to `otterdroid` as I have to create a new
Play Store listing as I no longer have access to the signing key.

More information can be found on Pjuu.com and the F-Droid forums.